% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-data.R
\name{add_point}
\alias{add_point}
\title{Create data for an additional point at a user-specified location.}
\usage{
add_point(feature)
}
\arguments{
\item{feature}{a leaflet point feature.}
}
\value{
a single-rowed tibble with the occurrence data in it.
}
\description{
Takes a leaflet feature created during user map-editing and wrangles the
data into the internal data format.
}
