% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projWiz.R
\name{cart2ll}
\alias{cart2ll}
\title{Cartesian (x,y,z) to Geographic coordinates}
\usage{
cart2ll(x, y, z)
}
\arguments{
\item{x}{East to West coordinate in metres}

\item{y}{South to North coordinate in metres}

\item{z}{height coordinate in metres}
}
\value{
dataframe of latitude,longtitude
}
\description{
calculates the latitude and longitude cordinates in radians from Cartesian coordinates (x,y,z). Used as part of the projection wizard.
}
\author{
Justin Moat. J.Moat@kew.org
}
