% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{constructPolygon}
\alias{constructPolygon}
\title{Construct a polygon from vertices.}
\usage{
constructPolygon(x, y, crs)
}
\description{
Accepts an x and a y vector to define the vertices of
the polygon, to make it easier
}
