% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projWiz.R
\name{deg2rad}
\alias{deg2rad}
\title{Degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{number in degrees}
}
\value{
number
}
\description{
Calculates radians from degrees
}
\examples{
a <- 30
deg2rad(a)
}
\author{
Justin Moat. J.Moat@kew.org
}
