% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif.R
\name{get_gbif_points}
\alias{get_gbif_points}
\title{Get a table of occurrence points from GBIF.}
\usage{
get_gbif_points(name, limit = 900)
}
\arguments{
\item{name}{scientific name of the taxon to request points for.}

\item{limit}{the maximum number of points to request.}
}
\value{
a standardised tibble of occurrence records.
}
\description{
Request points from GBIF and parses them into a standard format. By default,
gets the first 900 points with coordinates and no geospatial issues for a name.
If the exact name entered is not in the GBIF backbone, no points will be returned.
}
\examples{
points <- get_gbif_points("Poa annua")

}
