% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{import_csv}
\alias{import_csv}
\title{Import a table of occurrence points from a user-provided CSV.}
\usage{
import_csv(path)
}
\arguments{
\item{path}{path to csv to read.}
}
\value{
a standardised tibble of occurrence records
}
\description{
Load points from a CSV and parse them into a standard format. All columns
that don't match the standard format will be discard.
}
\examples{
points <- import_csv("path/to/file")

}
