% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{import_gbif}
\alias{import_gbif}
\title{Import a table of occurrence points from GBIF.}
\usage{
import_gbif(name, limit = 900)
}
\arguments{
\item{name}{scientific name of the taxon to request points for.}

\item{limit}{the maximum number of points to request.}
}
\value{
a standardised tibble of occurrence records.
}
\description{
Request points from GBIF and parses them into a standard format. By default,
gets the first 900 points with coordinates and no geospatial issues for a name.
If the exact name entered is not in the GBIF backbone, no points will be returned.
}
\examples{
points <- import_gbif("Poa annua")

}
