% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projWiz.R
\name{ll2cart}
\alias{ll2cart}
\title{Geographic coordinates to cartesian (x,y,z)}
\usage{
ll2cart(latr, longr)
}
\arguments{
\item{latr}{latitude point in radians}

\item{longr}{longtitude point in radians}
}
\value{
dataframe of x,y,z
}
\description{
Calculates the Cartesian coordinates (x,y,z) from lat long in radians. Used as part of the projection wizard.
}
\examples{
lat <- runif (200,-24,-12)
long <- runif (200,43,51)
thepoints <- data.frame(lat,long)
llrad <- deg2rad(thepoints)
cartp <- ll2cart(llrad$lat,llrad$long)
}
\references{
Descartes, R., 1637. Discours de la methode. A Leyde, De l’imprimerie de I. Maire, Paris.
}
\author{
Justin Moat. J.Moat@kew.org
}
