% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projWiz.R
\name{rad2deg}
\alias{rad2deg}
\title{Radians to Degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{number in radians}
}
\value{
number
}
\description{
Calculates degrees from radians
}
\examples{
b <- 0.392699
rad2deg(b)
}
\author{
Justin Moat. J.Moat@kew.org
}
