% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ratingAoo}
\alias{ratingAoo}
\title{IUCN rating based on AOO Area}
\usage{
ratingAoo(AOOArea, abb = TRUE)
}
\arguments{
\item{AOOArea}{Area in km2}

\item{abb}{abbreviation TRUE or FALSE , TRUE = 2 letter code, FALSE = full text (default = TRUE)}
}
\value{
Text one of CR, EN, VU, NT, LC or Critically Endangered, Endangered, Vulnerable, Near Threatened, Least Concern
}
\description{
Calculates IUCN rating based on Area of occupancy (AOO) in km2
}
\note{
Near Threatened is set at 3,000 km2, follow example in IUCN Guidelines version 11. 2014
}
\examples{
#will return "Endangered"
ratingAoo(25,FALSE)
#will return "EN"
ratingAoo(25,TRUE)
}
\references{
Bachman, S., Moat, J., Hill, A.W., de Torre, J., Scott, B., 2011. Supporting Red List threat assessments with GeoCAT: geospatial conservation assessment tool. Zookeys 126, 117–26. doi:10.3897/zookeys.150.2109

Moat, J., Bachman, S., n.d. GeoCAT Geospatial Conservation Assessment Tool \link{WWW Document}. URL http://geocat.kew.org/

IUCN, 2012. IUCN Red List Categories and Criteria, 2nd ed. IUCN, Gland, Switzerland.

IUCN Standards and Petitions Subcommittee, 2014. Guidelines for Using the IUCN Red List Categories and Criteria. Version 11.

IUCN Standards and Petitions Subcommittee, 2016. Guidelines for Using the IUCN Red List Categories and Criteria. Version 12.

Joppa, L.N., Butchart, S.H.M., Hoffmann, M., Bachman, S.P., Akçakaya, H.R., Moat, J.F., Böhm, M., Holland, R.A., Newton, A., Polidoro, B., Hughes, A., 2016. Impact of alternative metrics on estimates of extent of occurrence for extinction risk assessment. Conserv. Biol. 30, 362–370. doi:10.1111/cobi.12591
}
\seealso{
\code{\link{aoo}} for AOO calculations
}
\author{
Justin Moat. J.Moat@kew.org
}
