% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ratingEoo}
\alias{ratingEoo}
\title{IUCN rating based on EOO Area}
\usage{
ratingEoo(EOOArea, abb = TRUE)
}
\arguments{
\item{EOOArea}{Area in km2}

\item{abb}{abbreviation TRUE or FALSE , TRUE = 2 letter code, FALSE = full text (see value), default = TRUE}
}
\value{
Text
one of CR, EN, VU, NT, LC or Critically Endangered, Endangered, Vulnerable, Near Threatened, Least Concern
}
\description{
Calculates IUCN rating based on Extent of Occurrence (EOO) Area in km2
}
\details{
Calculates IUCN rating on EOO
}
\note{
Any negative values are assumed to be positive. Near Threatened is set at 30,000 km2, follow example in IUCN petition 2014
}
\examples{
ratingEoo(250,TRUE)
ratingEoo(250,FALSE)
}
\references{
Bachman, S., Moat, J., Hill, A.W., de Torre, J., Scott, B., 2011. Supporting Red List threat assessments with GeoCAT: geospatial conservation assessment tool. Zookeys 126, 117–26. doi:10.3897/zookeys.150.2109

Moat, J., Bachman, S., n.d. GeoCAT Geospatial Conservation Assessment Tool \link{WWW Document}. URL http://geocat.kew.org/

IUCN, 2012. IUCN RED LIST CATEGORIES AND CRITERIA, 2nd ed. IUCN, Gland, Switzerland.

IUCN Standards and Petitions Subcommittee, 2014. Guidelines for Using the IUCN Red List Categories and Criteria. Version 11.

IUCN Standards and Petitions Subcommittee, 2016. Guidelines for Using the IUCN Red List Categories and Criteria. Version 12.

Joppa, L.N., Butchart, S.H.M., Hoffmann, M., Bachman, S.P., Akçakaya, H.R., Moat, J.F., Böhm, M., Holland, R.A., Newton, A., Polidoro, B., Hughes, A., 2016. Impact of alternative metrics on estimates of extent of occurrence for extinction risk assessment. Conserv. Biol. 30, 362–370. doi:10.1111/cobi.12591
}
\seealso{
\code{\link{eoo}} for EOO calculations
}
\author{
Justin Moat. J.Moat@kew.org
}
