% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rotateP}
\alias{rotateP}
\title{Rotates a set of points}
\usage{
rotateP(thepoints, angle)
}
\arguments{
\item{thepoints}{set of points in metres i.e. c(x,y)}

\item{angle}{in radians}
}
\value{
dataframe of points
}
\description{
Rotates a set of point by an angle in radians. Used as part of the AOO rotation calculations.
}
\examples{
#Build and project some points
thepoints <- ptsSquare(200,0.1)
#rotate by 45 degrees
rpoints <- rotateP(thepoints,deg2rad(45))
plot(rpoints,asp=1)
}
\references{
Moat, J., Bachman, S. P., Field, R., & Boyd, D. S. (2018). Refining area of occupancy to address the modifiable areal unit problem in ecology and conservation. Conservation biology, 32(6), 1278-1289.
}
\author{
Justin Moat. J.Moat@kew.org
}
